Add files via upload
The aim of this project is to simulate the progression of Covid-19 in a chosen set of countries using Pandas
and NumPy. The Markov Chain rule is implemented on each timeline, and the condensed time series is
saved in ”.CSV” files. To conclude, the spread of Covid-19 is visualized and saved as a ”.PNG” file. This
project aims to showcase the benefit of using historical data to study the spread of the disease.